+++ spec/rubocop/cop/cop_spec.rb
@@ -244,7 +244,7 @@ describe RuboCop::Cop::Cop do
   describe '#style_guide_url' do
     let(:options) { {} }
     let(:cop) { described_class.new(config, options) }
-    subject { cop.style_guide_url }
+    subject(:url) { cop.style_guide_url }
 
     context 'when StyleGuide is not set in the config' do
       let(:config) { RuboCop::Config.new({}) }
@@ -258,6 +258,46 @@ describe RuboCop::Cop::Cop do
         )
       end
       it { is_expected.to eq('http://example.org/styleguide') }
+    end
+
+    context 'when a base URL is specified' do
+      let(:config) do
+        RuboCop::Config.new(
+          'AllCops' => {
+            'StyleGuideBaseURL' => 'http://example.org/styleguide'
+          }
+        )
+      end
+
+      it 'does not specify a URL if a cop does not have one' do
+        config['Cop/Cop'] = { 'StyleGuide' => nil }
+        expect(url).to be_nil
+      end
+
+      it 'combines correctly with a target-based setting' do
+        config['Cop/Cop'] = { 'StyleGuide' => '#target_based_url' }
+        expect(url).to eq('http://example.org/styleguide#target_based_url')
+      end
+
+      it 'can use a path-based setting' do
+        config['Cop/Cop'] = { 'StyleGuide' => 'cop/path/rule#target_based_url' }
+        expect(url).to eq('http://example.org/cop/path/rule#target_based_url')
+      end
+
+      it 'can accept relative paths if base has a full path' do
+        config['AllCops'] = {
+          'StyleGuideBaseURL' => 'http://github.com/bbatsov/ruby-style-guide/'
+        }
+        config['Cop/Cop'] = {
+          'StyleGuide' => '../rails-style-guide#target_based_url'
+        }
+        expect(url).to eq('http://github.com/bbatsov/rails-style-guide#target_based_url')
+      end
+
+      it 'allows absolute URLs in the cop config' do
+        config['Cop/Cop'] = { 'StyleGuide' => 'http://other.org#absolute_url' }
+        expect(url).to eq('http://other.org#absolute_url')
+      end
     end
   end
 end